movies_admin/requirements/dev.txt
django-debug-toolbar==3.2.2
django_extensions==3.1.3
ipython==7.27.0